+++ src/files/domain/models/FileCitationFormat.ts
@@ -1,0 +1,7 @@
+export enum FileCitationFormat {
+  ENDNOTE = 'EndNote',
+  RIS = 'RIS',
+  BIBTEX = 'BibTeX',
+  CSL = 'CSL',
+  INTERNAL = 'Internal'
+}